[MISC] move interface -> src/interface/Portfolio.ts
